+++ src/dotnet/APIView/APIViewWeb/Models/UploadModel.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
 using Microsoft.AspNetCore.Http;
@@ -13,5 +13,13 @@ namespace APIViewWeb.Pages.Assemblies
 
         [BindProperty]
         public IFormFile[] Files { get; set; }
+
+        [BindProperty]
+        public string Language { get; set; }
+
+        [BindProperty]
+        public string FilePath { get; set; }
+
+        public static string[] SupportedLanguages => LanguageService.SupportedLanguages;
     }
 }