+++ src/dotnet/APIView/APIViewWeb/Startup.cs
@@ -24,6 +24,7 @@ using APIViewWeb.HostedServices;
 using APIViewWeb.Filters;
 using APIViewWeb.Account;
 using APIView.Identity;
+using APIViewWeb.Managers;
 
 namespace APIViewWeb
 {
@@ -66,9 +67,13 @@ namespace APIViewWeb
                 .GetSection("Github")
                 .Bind(options));
 
+#pragma warning disable ASP5001 // Type or member is obsolete
+#pragma warning disable CS0618 // Type or member is obsolete
             services.AddMvc()
                 .SetCompatibilityVersion(CompatibilityVersion.Latest)
                 .AddRazorRuntimeCompilation();
+#pragma warning restore CS0618 // Type or member is obsolete
+#pragma warning restore ASP5001 // Type or member is obsolete
 
             services.AddRazorPages(options =>
             {
@@ -76,24 +81,25 @@ namespace APIViewWeb
                 options.Conventions.AddPageRoute("/Assemblies/Index", "");
             });
 
-            services.AddSingleton<BlobCodeFileRepository>();
-            services.AddSingleton<BlobOriginalsRepository>();
-            services.AddSingleton<CosmosReviewRepository>();
-            services.AddSingleton<CosmosCommentsRepository>();
-            services.AddSingleton<CosmosPullRequestsRepository>();
+            services.AddSingleton<IBlobCodeFileRepository, BlobCodeFileRepository>();
+            services.AddSingleton<IBlobOriginalsRepository, BlobOriginalsRepository>();
+            services.AddSingleton<IBlobUsageSampleRepository, BlobUsageSampleRepository>();
+            services.AddSingleton<ICosmosReviewRepository,CosmosReviewRepository>();
+            services.AddSingleton<ICosmosCommentsRepository, CosmosCommentsRepository>();
+            services.AddSingleton<ICosmosPullRequestsRepository, CosmosPullRequestsRepository>();
+            services.AddSingleton<ICosmosUsageSampleRepository, CosmosUsageSampleRepository>();
+            services.AddSingleton<ICosmosUserProfileRepository, CosmosUserProfileRepository>();
             services.AddSingleton<IDevopsArtifactRepository, DevopsArtifactRepository>();
-            services.AddSingleton<CosmosUsageSampleRepository>();
-            services.AddSingleton<BlobUsageSampleRepository>();
-            services.AddSingleton<CosmosUserProfileRepository>();
 
-            services.AddSingleton<ReviewManager>();
-            services.AddSingleton<CommentsManager>();
-            services.AddSingleton<NotificationManager>();
-            services.AddSingleton<PullRequestManager>();
-            services.AddSingleton<PackageNameManager>();
+            services.AddSingleton<IReviewManager, ReviewManager>();
+            services.AddSingleton<ICommentsManager, CommentsManager>();
+            services.AddSingleton<INotificationManager, NotificationManager>();
+            services.AddSingleton<IPullRequestManager, PullRequestManager>();
+            services.AddSingleton<IPackageNameManager, PackageNameManager>();
+            services.AddSingleton<IUsageSampleManager, UsageSampleManager>();
+            services.AddSingleton<IUserProfileManager, UserProfileManager>();
+            services.AddSingleton<IOpenSourceRequestManager, OpenSourceRequestManager>();
             services.AddSingleton<UserPreferenceCache>();
-            services.AddSingleton<UsageSampleManager>();
-            services.AddSingleton<UserProfileManager>();
 
             services.AddSingleton<LanguageService, JsonLanguageService>();
             services.AddSingleton<LanguageService, CSharpLanguageService>();
@@ -107,11 +113,13 @@ namespace APIViewWeb
             services.AddSingleton<LanguageService, SwaggerLanguageService>();
             services.AddSingleton<LanguageService, SwiftLanguageService>();
             services.AddSingleton<LanguageService, XmlLanguageService>();
+            services.AddSingleton<LanguageService, CadlLanguageService>();
 
             if (Environment.IsDevelopment() && Configuration["AuthenticationScheme"] == "Test")
             {
                 services.AddAuthentication("Test")
                     .AddScheme<AuthenticationSchemeOptions, TestAuthHandler>("Test", options => { });
+                services.AddSingleton<IStartupFilter, UITestsStartUpFilter>();
             }
             else
             {
@@ -209,7 +217,6 @@ namespace APIViewWeb
             services.AddSingleton<IAuthorizationHandler, ApproverRequirementHandler>();
             services.AddSingleton<IAuthorizationHandler, ResolverRequirementHandler>();
             services.AddSingleton<IAuthorizationHandler, AutoReviewModifierRequirementHandler>();
-            services.AddSingleton<IAuthorizationHandler, PullRequestPermissionRequirementHandler>();
             services.AddSingleton<IAuthorizationHandler, UsageSampleOwnerRequirementHandler>();
             services.AddHostedService<ReviewBackgroundHostedService>();
             services.AddHostedService<PullRequestBackgroundHostedService>();