+++ tools/apiview/emitters/cadl-apiview/src/navigation.ts
@@ -1,4 +1,5 @@
 import {
+  AliasStatementNode,
   EnumStatementNode,
   InterfaceStatementNode,
   IntersectionExpressionNode,
@@ -6,11 +7,12 @@ import {
   ModelStatementNode,
   OperationStatementNode,
   ProjectionModelExpressionNode,
+  ScalarStatementNode,
   SyntaxKind,
   UnionExpressionNode,
   UnionStatementNode,
 } from "@cadl-lang/compiler";
-import { NamespaceStack } from "./apiview.js";
+import { ApiView, NamespaceStack } from "./apiview.js";
 import { NamespaceModel } from "./namespace-model.js";
 
 export class ApiViewNavigation {
@@ -21,6 +23,7 @@ export class ApiViewNavigation {
 
   constructor(
     objNode:
+      | AliasStatementNode
       | NamespaceModel
       | ModelStatementNode
       | OperationStatementNode
@@ -29,6 +32,7 @@ export class ApiViewNavigation {
       | ModelExpressionNode
       | IntersectionExpressionNode
       | ProjectionModelExpressionNode
+      | ScalarStatementNode
       | UnionStatementNode
       | UnionExpressionNode,
       stack: NamespaceStack
@@ -51,6 +55,10 @@ export class ApiViewNavigation {
         for (const node of objNode.models.values()) {
           modelItems.push(new ApiViewNavigation(node, stack));
         }
+        const aliasItems = new Array<ApiViewNavigation>();
+        for (const node of objNode.aliases.values()) {
+            aliasItems.push(new ApiViewNavigation(node, stack));
+        }
         this.ChildItems = [];
         if (operationItems.length) {
           this.ChildItems.push({ Text: "Operations", ChildItems: operationItems, Tags: { TypeKind: ApiViewNavigationKind.Method }, NavigationId: "" });
@@ -60,6 +68,9 @@ export class ApiViewNavigation {
         }
         if (modelItems.length) {
           this.ChildItems.push({ Text: "Models", ChildItems: modelItems, Tags: { TypeKind: ApiViewNavigationKind.Class }, NavigationId: "" });
+        }
+        if (aliasItems.length) {
+            this.ChildItems.push({ Text: "Aliases", ChildItems: aliasItems, Tags: { TypeKind: ApiViewNavigationKind.Class }, NavigationId: "" });
         }
         break;
       case SyntaxKind.ModelStatement:
@@ -100,12 +111,26 @@ export class ApiViewNavigation {
         this.Tags = { TypeKind: ApiViewNavigationKind.Enum };
         this.ChildItems = [];
         break;
+      case SyntaxKind.AliasStatement:
+        obj = objNode as AliasStatementNode;
+        stack.push(obj.id.sv);
+        this.Text = obj.id.sv;
+        this.Tags = { TypeKind: ApiViewNavigationKind.Class };
+        this.ChildItems = [];
+        break;
       case SyntaxKind.ModelExpression:
         throw new Error(`Navigation unsupported for "ModelExpression".`);
       case SyntaxKind.IntersectionExpression:
         throw new Error(`Navigation unsupported for "IntersectionExpression".`);
       case SyntaxKind.ProjectionModelExpression:
         throw new Error(`Navigation unsupported for "ProjectionModelExpression".`);
+      case SyntaxKind.ScalarStatement:
+        obj = objNode as ScalarStatementNode;
+        stack.push(obj.id.sv);
+        this.Text = obj.id.sv;
+        this.Tags = { TypeKind: ApiViewNavigationKind.Class };
+        this.ChildItems = [];
+        break;
       default:
         throw new Error(`Navigation unsupported for "${objNode.kind.toString()}".`);
     }