+++ tools/apiview/parsers/swagger-api-parser/SwaggerApiParser/SwaggerSpec/SwaggerTypes.cs
@@ -164,7 +164,6 @@ namespace SwaggerApiParser
 
             if (schema.allOfProperities?.Count != 0 && schema.allOf != null)
             {
-                // ret.Add(TokenSerializer.Intent(context.intent));
                 ret.Add(new CodeFileToken("allOf", CodeFileTokenKind.Keyword));
                 ret.Add(TokenSerializer.Colon());
                 ret.Add(TokenSerializer.NewLine());
@@ -234,7 +233,6 @@ namespace SwaggerApiParser
 
             foreach (var kv in properties)
             {
-                // ret.Add(TokenSerializer.Intent(context.intent));
                 ret.Add(new CodeFileToken(kv.Key, CodeFileTokenKind.Literal));
                 ret.Add(TokenSerializer.Colon());
                 if (kv.Value == null)
@@ -258,7 +256,6 @@ namespace SwaggerApiParser
                 else if (kv.Value.Ref != null)
                 {
                     ret.Add(TokenSerializer.NewLine());
-                    // ret.Add(TokenSerializer.Intent(context.intent + 1));
                     ret.Add(new CodeFileToken("<", CodeFileTokenKind.Punctuation));
                     var refName = kv.Value.Ref;
                     ret.Add(new CodeFileToken(refName.Split("/").Last(), CodeFileTokenKind.TypeName));
@@ -308,7 +305,17 @@ namespace SwaggerApiParser
                 ret.Add(new CodeFileToken(refName.Split("/").Last(), CodeFileTokenKind.TypeName));
                 ret.Add(new CodeFileToken(">", CodeFileTokenKind.Punctuation));
                 ret.Add(TokenSerializer.NewLine());
-                ret.AddRange(arraySchema.items.TokenSerializeInternal(new SerializeContext(context.intent + 1, context.IteratorPath), arraySchema.items, ref flattenedTableItems, serializeRef));
+                
+                // circular reference
+                if (arraySchema.items.Ref != null)
+                {
+                    return;
+                }
+
+                if (serializeRef)
+                {
+                    ret.AddRange(arraySchema.items.TokenSerializeInternal(new SerializeContext(context.intent + 1, context.IteratorPath), arraySchema.items, ref flattenedTableItems, serializeRef));
+                }
             }
         }
 
@@ -319,7 +326,6 @@ namespace SwaggerApiParser
                 retTableItems = new List<SchemaTableItem>();
                 this.TokenSerializeInternal(context, this, ref retTableItems, serializeRef);
             }
-            // var ret = this.TokenSerializeInternal(context, this, ref flattenedSchema);
         }
 
         public CodeFileToken[] TokenSerialize(SerializeContext context)