+++ tools/apiview/emitters/cadl-apiview/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## Version 0.3.3 (01-03-2023)
+Fixed issue where some type references were not navigable.
+
+## Version 0.3.2 (12-20-2022)
+Changed structure of APIView navigation so that aliases appear under a separate "Alias" section, instead of
+  within the existing "Models" section. Will likely result in a non-API-related diff with prior APIView versions.
+
+## Version 0.3.1 (12-9-2022)
+Support Cadl scalars.
+
+## Version 0.3.0 (11-15-2022)
+Add support for aliases and augment decorators.
+
 ## Version 0.2.1 (10-27-2022)
 Change behavior of `version` emitter option so that if it is not supplied, APIView will be generated for the
   un-projected Cadl, rendering all versioning decorators. Supplying `version` allows the user to project a